+++ book/book-examples/tests/view-example-string.rs
@@ -2,7 +2,7 @@ use bonsaidb::core::connection::Connection;
 use bonsaidb::core::document::{BorrowedDocument, Emit};
 use bonsaidb::core::schema::view::map::ViewMappedValue;
 use bonsaidb::core::schema::{
-    Collection, ReduceResult, SerializedCollection, View, ViewMapResult, ViewSchema,
+    Collection, MapReduce, ReduceResult, SerializedCollection, View, ViewMapResult, ViewSchema,
 };
 use bonsaidb::core::Error;
 use bonsaidb::local::config::{Builder, StorageConfiguration};
@@ -20,14 +20,11 @@ pub struct BlogPost {
 // ANCHOR_END: struct
 
 // ANCHOR: view
-#[derive(Debug, Clone, View)]
+#[derive(Debug, Clone, View, ViewSchema)]
 #[view(collection = BlogPost, key = Option<String>, value = u32, name = "by-category")]
 pub struct BlogPostsByCategory;
 
-impl ViewSchema for BlogPostsByCategory {
-    type MappedKey<'doc> = Option<String>;
-    type View = Self;
-
+impl MapReduce for BlogPostsByCategory {
     fn map<'doc>(&self, document: &'doc BorrowedDocument<'_>) -> ViewMapResult<'doc, Self> {
         let post = BlogPost::document_contents(document)?;
         document.header.emit_key_and_value(post.category, 1)