+++ crates/bonsaidb-core/src/schema/schematic.rs
@@ -9,7 +9,9 @@ use crate::document::{BorrowedDocument, DocumentId, KeyId};
 use crate::key::{ByteSource, Key, KeyDescription};
 use crate::schema::collection::Collection;
 use crate::schema::view::map::{self, MappedValue};
-use crate::schema::view::{self, Serialized, SerializedView, ViewSchema};
+use crate::schema::view::{
+    self, MapReduce, Serialized, SerializedView, ViewSchema, ViewUpdatePolicy,
+};
 use crate::schema::{CollectionName, Schema, SchemaName, View, ViewName};
 use crate::Error;
 
@@ -66,7 +68,7 @@ impl Schematic {
     }
 
     /// Adds the view `V`.
-    pub fn define_view<V: ViewSchema<View = V> + SerializedView + Clone + 'static>(
+    pub fn define_view<V: MapReduce + ViewSchema<View = V> + SerializedView + Clone + 'static>(
         &mut self,
         view: V,
     ) -> Result<(), Error> {
@@ -76,7 +78,7 @@ impl Schematic {
     /// Adds the view `V`.
     pub fn define_view_with_schema<
         V: SerializedView + 'static,
-        S: ViewSchema<View = V> + 'static,
+        S: MapReduce + ViewSchema<View = V> + 'static,
     >(
         &mut self,
         view: V,
@@ -89,7 +91,7 @@ impl Schematic {
         }
 
         let collection = instance.collection();
-        let eager = instance.eager();
+        let eager = instance.update_policy().is_eager();
         self.views.insert(TypeId::of::<V>(), Box::new(instance));
         self.views_by_name.insert(name, TypeId::of::<V>());
 
@@ -213,7 +215,7 @@ struct ViewInstance<V, S> {
 impl<V, S> Serialized for ViewInstance<V, S>
 where
     V: SerializedView,
-    S: ViewSchema<View = V>,
+    S: MapReduce + ViewSchema<View = V>,
 {
     fn collection(&self) -> CollectionName {
         <<V as View>::Collection as Collection>::collection_name()
@@ -223,12 +225,8 @@ where
         KeyDescription::for_key::<<V as View>::Key>()
     }
 
-    fn unique(&self) -> bool {
-        self.schema.unique()
-    }
-
-    fn lazy(&self) -> bool {
-        self.schema.lazy()
+    fn update_policy(&self) -> ViewUpdatePolicy {
+        self.schema.update_policy()
     }
 
     fn version(&self) -> u64 {