+++ crates/bonsaidb-macros/tests/ui/view_schema/invalid_attribute.rs
@@ -1,0 +1,7 @@
+use bonsaidb::core::schema::ViewSchema;
+
+#[derive(ViewSchema, Debug)]
+#[view_schema(hi)]
+struct Test1;
+
+fn main() {}